Move main back to /app
