+++ app/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/echo-webkom/cenv/cmd"
+
+func main() {
+	cmd.Run()
+}
